+++ examples/area_gallery.rb
@@ -42,7 +42,7 @@ window('Area Gallery', 400, 400) {
       stroke r: 0, g: 0, b: 0, thickness: 2, dashes: [50, 10, 10, 10], dash_phase: -50.0
     }
     path { # declarative stable path
-      arc(400, 220, 180, 90, 90, false)
+      arc(404, 216, 190, 90, 90, false)
 
       # radial gradient (has an outer_radius in addition to x0, y0, x1, y1, and stops)
       fill outer_radius: 90, x0: 0, y0: 0, x1: 500, y1: 500, stops: [{pos: 0.25, r: 102, g: 102, b: 204, a: 0.5}, {pos: 0.75, r: 204, g: 102, b: 204}]
@@ -54,7 +54,23 @@ window('Area Gallery', 400, 400) {
       fill r: 202, g: 102, b: 204, a: 0.5
       stroke r: 0, g: 0, b: 0, thickness: 2
     }
-    text(160, 40, 100) { # x, y, width
+    path { # declarative stable path
+      polygon(370, 150, 325, 275, 325, 125, 370, 250)
+
+      fill r: 202, g: 102, b: 204, a: 0.5
+      stroke r: 0, g: 0, b: 0, thickness: 2
+    }
+    path { # declarative stable path
+      polyline(150, 50, 175, 25, 225, 25, 250, 50)
+
+      stroke r: 0, g: 0, b: 0, thickness: 2
+    }
+    path { # declarative stable path
+      polybezier(212, 212, 125, 175, 150, 350, 300, 300, 450, 250, 350, 250, 400, 400)
+
+      stroke r: 0, g: 0, b: 0, thickness: 2
+    }
+    text(164, 40, 100) { # x, y, width
       string('Area Gallery') {
         font family: 'Times', size: 14
         color :black